[reg_ports] start bench for reg file ports
Currently just trying to predictably get nvcc to allocate all the
registers in a SM partition. Since we are invoking the kernel in a
single thread block, all the threads should get swapped in and out of a
single partition. Right now having trouble getting the compiler to
predictably use registers over shared memory, etc.

8 registers seem to be used for basic kernel state required by the
compiler and are register overheads on top of what is required by the
kernel code itself. Registers are only allocated for kernel local
variables, but if local arrays get too large, the compiler moves them.